Breaks up Apache config tests into multiple files
The combined test file was approaching 400 lines, much of which was
variable declaration via the pytest parametrization marker. Became
difficult to navigate and update. Splitting up the tests into smaller
chunks makes it more obvious where testing logic should land.
